일단 다리 4개 완성 아직 튼튼할 때는 해야함!
